+++ hyperbolix/__init__.py
@@ -1,8 +1,11 @@
 """Hyperbolix JAX backend - pure functional implementation."""
 
 from . import manifolds, utils
+from .utils.curvature import get_curvature, learnable_curvature
 
 __all__ = [
+    "get_curvature",
+    "learnable_curvature",
     "manifolds",
     "utils",
 ]