+++ hyperbolix/utils/__init__.py
@@ -1,5 +1,6 @@
 """JAX utilities for hyperbolix."""
 
+from .curvature import get_curvature, learnable_curvature
 from .helpers import compute_hyperbolic_delta, compute_pairwise_distances, get_delta
 from .math_utils import acosh, atanh, cosh, sinh, smooth_clamp, smooth_clamp_max, smooth_clamp_min
 
@@ -9,7 +10,9 @@ __all__ = [
     "compute_hyperbolic_delta",
     "compute_pairwise_distances",
     "cosh",
+    "get_curvature",
     "get_delta",
+    "learnable_curvature",
     "sinh",
     "smooth_clamp",
     "smooth_clamp_max",